+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/AnnotationHandlerAdapter.java
@@ -1,6 +1,6 @@
 package com.interface21.webmvc.servlet.mvc.tobe;
 
-import com.interface21.webmvc.servlet.HandlerAdapter;
+import com.interface21.webmvc.servlet.adapter.HandlerAdapter;
 import com.interface21.webmvc.servlet.ModelAndView;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
